Refactored code using encapsulation: menu implementation is now singleton class (object).Jira task: FIT-72
