+++ Android/app/src/main/java/hr/foi/air/fitfusion/WelcomeTrainerActivity.kt
@@ -12,6 +12,7 @@ import com.google.android.material.tabs.TabLayoutMediator
 import hr.foi.air.fitfusion.adapters.MainPagerAdapter
 import hr.foi.air.fitfusion.data_classes.LoggedInUser
 import hr.foi.air.fitfusion.databinding.ActivityTrainerWelcomeBinding
+import hr.foi.air.fitfusion.entities.ProfileMenu
 import hr.foi.air.fitfusion.fragments.CalendarFragment
 import hr.foi.air.fitfusion.fragments.ForumFragment
 import hr.foi.air.fitfusion.fragments.HomeTrainerFragment
@@ -34,11 +35,33 @@ class WelcomeTrainerActivity : AppCompatActivity() {
 
         val imgButton = findViewById<ImageButton>(R.id.imageButtonUser)
 
-        imgButton.setOnClickListener(){
+        /*imgButton.setOnClickListener(){
             val popup = PopupMenu(this, it)
             val inflater: MenuInflater = popup.menuInflater
             inflater.inflate(R.menu.account_menu, popup.menu)
             popup.show()
+        }*/
+
+        imgButton.setOnClickListener(){ button ->
+            ProfileMenu.showMenu(
+                context = this,
+                anchor = button,
+                menuRes = R.menu.account_menu,
+                actionHandler = { itemId ->
+                    when (itemId) {
+                        R.id.settings_option -> {
+                            //take user to settings
+                        }
+                        R.id.logout_option -> {
+                            //handle logout action
+                        }
+                        else -> {
+                            //something else
+                        }
+                    }
+                }
+            )
+
         }
 
         val mainPagerAdapter = MainPagerAdapter(supportFragmentManager, lifecycle)
